requirements.txt
#python_lib
#2019-01-15
kombu==4.2.1
librabbitmq==2.0.0
cffi==1.11.5
pika==0.12.0
gevent==1.3.7
six==1.12.0
mmh3==2.5.1
PyYAML==3.13
mock==2.0.0
pytest==4.0.2
pytest-cov==2.6.0
pylint==1.9.4
requests==2.21.0
setuptools==40.6.3
ipaddr==2.2.0
redis==2.10.6
lxml==4.2.5
chardet==3.0.4
sqlalchemy==1.2.15
influxdb==5.2.1
thrift==0.11.0
ipcalc==1.99.0
redis-py-cluster==1.3.6
raven==6.10.0
IPy==0.83
jinja2==2.10
pylru==1.1.0
plyvel==1.0.5
uwsgi==2.0.17.1

